% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{compare}
\alias{compare}
\title{Compare two players}
\usage{
compare(handle1, handle2 = "BaumerB", ...)
}
\arguments{
\item{handle1}{Opponent's player handle}

\item{handle2}{Your player handle}

\item{...}{currently ignored}
}
\description{
Compare two players
}
