% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_player_data}
\alias{scrape_player_data}
\title{Scrape player info}
\usage{
scrape_player_data(handle = "BaumerB", ...)
}
\arguments{
\item{handle}{A player's LearnedLeague handle}

\item{...}{currently ignored}
}
\description{
Scrape player info
}
\examples{

\dontrun{
x <- scrape_player_data()
}
}
